refactor - Sections 일급컬렉션 추가
