+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,16 +1,9 @@
 package nextstep.subway.line.domain;
 
 import nextstep.subway.BaseEntity;
-import nextstep.subway.constant.ErrorMessage;
-import nextstep.subway.exception.BadRequestException;
 import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.domain.Stations;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
 
 @Entity
 public class Line extends BaseEntity {
@@ -21,20 +14,23 @@ public class Line extends BaseEntity {
     private String name;
     private String color;
 
-    @OneToMany(mappedBy = "line", cascade = {CascadeType.PERSIST, CascadeType.MERGE}, orphanRemoval = true)
-    private List<Section> sections = new ArrayList<>();
+    @Embedded
+    private Sections sections;
 
     public Line() {
+        this.sections = new Sections();
     }
 
     public Line(String name, String color) {
         this.name = name;
         this.color = color;
+        this.sections = new Sections();
     }
 
     public Line(String name, String color, Station upStation, Station downStation, int distance) {
         this.name = name;
         this.color = color;
+        this.sections = new Sections();
         sections.add(new Section(this, upStation, downStation, new Distance(distance)));
     }
 
@@ -55,124 +51,7 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Section> getSections() {
+    public Sections getSections() {
         return sections;
-    }
-
-    public List<Station> getStations() {
-        if (sections.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        return findStations();
-    }
-
-    public void addLineStation(Section section) {
-        Stations stations = new Stations(getStations());
-        boolean isUpStationExisted = stations.isContainStations(section.getUpStation());
-        boolean isDownStationExisted = stations.isContainStations(section.getDownStation());
-
-        validateAddLineStations(stations, isUpStationExisted, isDownStationExisted);
-
-        addSections(section.getUpStation(), section.getDownStation(), section.getDistance(), stations, isUpStationExisted, isDownStationExisted);
-    }
-
-    private void validateAddLineStations(Stations stations, boolean isUpStationExisted, boolean isDownStationExisted) {
-        if (isUpStationExisted && isDownStationExisted) {
-            throw new BadRequestException(ErrorMessage.ALREADY_EXIST_SECTION);
-        }
-
-        if (!stations.isEmpty() && !isUpStationExisted && !isDownStationExisted) {
-            throw new BadRequestException(ErrorMessage.CAN_NOT_ADD_SECTION);
-        }
-
-        if (!isUpStationExisted && !isDownStationExisted) {
-            throw new BadRequestException(ErrorMessage.NOT_FOUND_SECTION_STATION);
-        }
-    }
-
-    private void addSections(Station upStation, Station downStation, Distance distance, Stations stations, boolean isUpStationExisted, boolean isDownStationExisted) {
-        if (stations.isEmpty()) {
-            addSection(upStation, downStation, distance);
-            return;
-        }
-
-        if (isUpStationExisted) {
-            sections.stream()
-                    .filter(it -> it.getUpStation() == upStation)
-                    .findFirst()
-                    .ifPresent(it -> it.updateUpStation(downStation, distance));
-
-            addSection(upStation, downStation, distance);
-        } else if (isDownStationExisted) {
-            sections.stream()
-                    .filter(it -> it.getDownStation() == downStation)
-                    .findFirst()
-                    .ifPresent(it -> it.updateDownStation(upStation, distance));
-
-            addSection(upStation, downStation, distance);
-        }
-    }
-
-    private void addSection(Station upStation, Station downStation, Distance distance) {
-        sections.add(new Section(this, upStation, downStation, distance));
-    }
-
-    private List<Station> findStations() {
-        List<Station> stations = new ArrayList<>();
-        Station downStation = findUpStation();
-        stations.add(downStation);
-
-        while (Optional.ofNullable(downStation).isPresent()) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getUpStation() == finalDownStation)
-                    .findFirst();
-
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-
-            downStation = nextLineStation.get().getDownStation();
-            stations.add(downStation);
-        }
-        return stations;
-    }
-
-    private Station findUpStation() {
-        Station downStation = sections.get(0).getUpStation();
-
-        while (downStation != null) {
-            Station finalDownStation = downStation;
-            Optional<Section> nextLineStation = sections.stream()
-                    .filter(it -> it.getDownStation() == finalDownStation)
-                    .findFirst();
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-
-            downStation = nextLineStation.get().getUpStation();
-        }
-
-        return downStation;
-    }
-
-    public void removeLineStation(Station station) {
-        Optional<Section> upLineStation = sections.stream()
-                .filter(it -> it.getUpStation() == station)
-                .findFirst();
-        Optional<Section> downLineStation = sections.stream()
-                .filter(it -> it.getDownStation() == station)
-                .findFirst();
-
-        upLineStation.ifPresent(it -> sections.remove(it));
-        downLineStation.ifPresent(it -> sections.remove(it));
-
-        if (upLineStation.isPresent() && downLineStation.isPresent()) {
-            Station newUpStation = downLineStation.get().getUpStation();
-            Station newDownStation = upLineStation.get().getDownStation();
-            int newDistance = upLineStation.get().getDistance().toInt() + downLineStation.get().getDistance().toInt();
-            addSection(newUpStation, newDownStation, new Distance(newDistance));
-        }
     }
 }
